WIP: some more restructurings and checking data interpretation.

+++ rpg-game-modules/rpg-torg/src/main/java/de/kaiserpfalzedv/rpg/torg/foundry/model/Details.java
@@ -37,7 +37,7 @@ import lombok.*;
 public class Details {
     private String possibilitypotential;
     private String description;
-    private float sizeBonus;
+    private int sizeBonus;
 
     // Archetype data
     private String clearance;
